Revert library name change
Libraries are locked to the `name` value specified by the library.properties metadata file in the release at the time of
the Library Manager submission.

Any release with a different name is rejected by the Arduino Library Manager indexer. The previous name change caused
the 1.6.0 release of the library to be rejected.

https://github.com/arduino/library-registry/blob/main/FAQ.md#update-requirements

> The name property in library.properties must not have changed from the value it had when the library was submitted.