Changes to Speaker Example Sketch
This fixes the noisy tone issue #193
Tone interrupt collides with our interrupt, noisy tone is the result, and likely bad BPM values.

In order to fix this problem and use tone() the Arduino core needs to be updated so that the tone library operates 'hands free' like the PWM library. Or, the PWM library needs to be updated to accept a frequency parameter to ensure a clean(er) note. There are some architectures upon which the Tone library might work: nRF52? ESP32?